LayoutTests/platform/wpe/fast/canvas/webgl/webgl-clear-composited-notshowing-expected.txt

Test that test that WebGL canvas preserveDrawingBuffer = false clear behavior on cases where compositor does not neccessarily display the element. You should see no failures and some of the canvases as green on screen.
PASS canvas: ref attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 0
PASS canvas: ref attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 0
PASS canvas: ref attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 0
PASS canvas: ref attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 0
PASS canvas: clipped attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 0
PASS canvas: clipped attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 0
PASS canvas: clipped attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 0
PASS canvas: clipped attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 0
PASS canvas: notOnScreen attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 0
PASS canvas: notOnScreen attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 0
PASS canvas: notOnScreen attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 0
PASS canvas: notOnScreen attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 0
PASS canvas: displayToNone attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 0
PASS canvas: displayToNone attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 0
PASS canvas: displayToNone attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 0
PASS canvas: displayToNone attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 0
PASS canvas: ref attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 1
PASS canvas: ref attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 1
PASS canvas: ref attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 1
PASS canvas: ref attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 1
PASS canvas: clipped attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 1
PASS canvas: clipped attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 1
PASS canvas: clipped attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 1
PASS canvas: clipped attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 1
PASS canvas: notOnScreen attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 1
PASS canvas: notOnScreen attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 1
PASS canvas: notOnScreen attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 1
PASS canvas: notOnScreen attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 1
FAIL canvas: displayToNone attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 1
at (0, 0) expected: 0,0,0,255 was 255,178,128,255
FAIL canvas: displayToNone attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 1
at (0, 0) expected: 0,0,0,0 was 255,178,128,255
FAIL canvas: displayToNone attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 1
at (0, 0) expected: 0,0,0,255 was 255,178,128,255
FAIL canvas: displayToNone attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 1
at (0, 0) expected: 0,0,0,0 was 255,178,128,255
PASS canvas: ref attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 2
PASS canvas: ref attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 2
PASS canvas: ref attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 2
PASS canvas: ref attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 2
PASS canvas: clipped attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 2
PASS canvas: clipped attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 2
PASS canvas: clipped attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 2
PASS canvas: clipped attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 2
PASS canvas: notOnScreen attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 2
PASS canvas: notOnScreen attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 2
PASS canvas: notOnScreen attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 2
PASS canvas: notOnScreen attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 2
PASS canvas: displayToNone attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 2
PASS canvas: displayToNone attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 2
PASS canvas: displayToNone attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 2
PASS canvas: displayToNone attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 2
PASS canvas: ref attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 3
PASS canvas: ref attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 3
PASS canvas: ref attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 3
PASS canvas: ref attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 3
PASS canvas: clipped attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 3
PASS canvas: clipped attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 3
PASS canvas: clipped attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 3
PASS canvas: clipped attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 3
PASS canvas: notOnScreen attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 3
PASS canvas: notOnScreen attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 3
PASS canvas: notOnScreen attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 3
PASS canvas: notOnScreen attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 3
FAIL canvas: displayToNone attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 3
at (0, 0) expected: 0,0,0,255 was 255,178,128,255
FAIL canvas: displayToNone attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 3
at (0, 0) expected: 0,0,0,0 was 255,178,128,255
FAIL canvas: displayToNone attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 3
at (0, 0) expected: 0,0,0,255 was 255,178,128,255
FAIL canvas: displayToNone attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 3
at (0, 0) expected: 0,0,0,0 was 255,178,128,255
PASS canvas: ref attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 4
PASS canvas: ref attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 4
PASS canvas: ref attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 4
PASS canvas: ref attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 4
PASS canvas: clipped attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 4
PASS canvas: clipped attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 4
PASS canvas: clipped attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 4
PASS canvas: clipped attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 4
PASS canvas: notOnScreen attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 4
PASS canvas: notOnScreen attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 4
PASS canvas: notOnScreen attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 4
PASS canvas: notOnScreen attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 4
PASS canvas: displayToNone attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 4
PASS canvas: displayToNone attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 4
PASS canvas: displayToNone attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 4
PASS canvas: displayToNone attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 4
PASS canvas: ref attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 5
PASS canvas: ref attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 5
PASS canvas: ref attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 5
PASS canvas: ref attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 5
PASS canvas: clipped attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 5
PASS canvas: clipped attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 5
PASS canvas: clipped attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 5
PASS canvas: clipped attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 5
PASS canvas: notOnScreen attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 5
PASS canvas: notOnScreen attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 5
PASS canvas: notOnScreen attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 5
PASS canvas: notOnScreen attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 5
FAIL canvas: displayToNone attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 5
at (0, 0) expected: 0,0,0,255 was 255,178,128,255
FAIL canvas: displayToNone attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 5
at (0, 0) expected: 0,0,0,0 was 255,178,128,255
FAIL canvas: displayToNone attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 5
at (0, 0) expected: 0,0,0,255 was 255,178,128,255
FAIL canvas: displayToNone attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 5
at (0, 0) expected: 0,0,0,0 was 255,178,128,255
PASS canvas: ref attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 6
PASS canvas: ref attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 6
PASS canvas: ref attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 6
PASS canvas: ref attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 6
PASS canvas: clipped attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 6
PASS canvas: clipped attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 6
PASS canvas: clipped attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 6
PASS canvas: clipped attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 6
PASS canvas: notOnScreen attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 6
PASS canvas: notOnScreen attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 6
PASS canvas: notOnScreen attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 6
PASS canvas: notOnScreen attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 6
PASS canvas: displayToNone attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 6
PASS canvas: displayToNone attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 6
PASS canvas: displayToNone attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 6
PASS canvas: displayToNone attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 6
PASS canvas: ref attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 7
PASS canvas: ref attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 7
PASS canvas: ref attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 7
PASS canvas: ref attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 7
PASS canvas: clipped attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 7
PASS canvas: clipped attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 7
PASS canvas: clipped attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 7
PASS canvas: clipped attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 7
PASS canvas: notOnScreen attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 7
PASS canvas: notOnScreen attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 7
PASS canvas: notOnScreen attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 7
PASS canvas: notOnScreen attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 7
FAIL canvas: displayToNone attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 7
at (0, 0) expected: 0,0,0,255 was 255,178,128,255
FAIL canvas: displayToNone attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 7
at (0, 0) expected: 0,0,0,0 was 255,178,128,255
FAIL canvas: displayToNone attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 7
at (0, 0) expected: 0,0,0,255 was 255,178,128,255
FAIL canvas: displayToNone attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 7
at (0, 0) expected: 0,0,0,0 was 255,178,128,255
PASS canvas: ref attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 8
PASS canvas: ref attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 8
PASS canvas: ref attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 8
PASS canvas: ref attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 8
PASS canvas: clipped attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 8
PASS canvas: clipped attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 8
PASS canvas: clipped attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 8
PASS canvas: clipped attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 8
PASS canvas: notOnScreen attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 8
PASS canvas: notOnScreen attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 8
PASS canvas: notOnScreen attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 8
PASS canvas: notOnScreen attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 8
PASS canvas: displayToNone attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 8
PASS canvas: displayToNone attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 8
PASS canvas: displayToNone attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 8
PASS canvas: displayToNone attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 8
PASS canvas: ref attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 9
PASS canvas: ref attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 9
PASS canvas: ref attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 9
PASS canvas: ref attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 9
PASS canvas: clipped attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 9
PASS canvas: clipped attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 9
PASS canvas: clipped attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 9
PASS canvas: clipped attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 9
PASS canvas: notOnScreen attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 9
PASS canvas: notOnScreen attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 9
PASS canvas: notOnScreen attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 9
PASS canvas: notOnScreen attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 9
FAIL canvas: displayToNone attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 9
at (0, 0) expected: 0,0,0,255 was 255,178,128,255
FAIL canvas: displayToNone attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 9
at (0, 0) expected: 0,0,0,0 was 255,178,128,255
FAIL canvas: displayToNone attrs: {"alpha":false,"antialias":false} is initialized to zero on iteration: 9
at (0, 0) expected: 0,0,0,255 was 255,178,128,255
FAIL canvas: displayToNone attrs: {"alpha":true,"antialias":false} is initialized to zero on iteration: 9
at (0, 0) expected: 0,0,0,0 was 255,178,128,255
PASS successfullyParsed is true
Some tests failed.

TEST COMPLETE
